+++ lua/spotify/config.lua
@@ -1,0 +1,13 @@
+local config = {
+  refresh_interval = 50,
+  cycle_speed = 5,
+  show_status_after_action = true,
+  timeout = 3000,
+  render = {
+    -- template
+    -- symbols
+    -- progress_bar_width
+    -- width
+  },
+}
+return config
